# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Oficina Mecanica - API (monolito)
  description: |
    API de gerenciamento da oficina: ordens de servico, catalogos, estoque e
    clientes.

    Autenticacao NAO faz parte deste contrato. `POST /auth/login` e
    `POST /auth/register` sao servidos por uma funcao Lambda cujo contrato vive
    em `oficina-mecanica-serverless/docs/openapi.yaml`. Este servico apenas
    valida o token emitido la; obtenha um em `/api/auth/login` antes de chamar
    qualquer rota protegida.
  version: 3.0.0
  license:
    name: Uso academico
servers:
  - url: /api
    description: Atras do CloudFront/API Gateway
  - url: /
    description: Servidor atual (execucao local direta)
...
security:
  - BearerAuth: []
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [Users]
      summary: Listar todos os usuarios
      description: Requer role admin.
      responses:
        <<: *protectedResponses
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getUsersId
      tags: [Users]
      summary: Buscar usuario por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putUsersId
      tags: [Users]
      summary: Atualizar usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Usuario atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteUsersId
      tags: [Users]
      summary: Deletar usuario
      responses:
        <<: *protectedResponses
        "204":
          description: Usuario deletado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Customers ─────────────────────────────────────────
  /customers:
    post:
      operationId: postCustomers
      tags: [Customers]
      summary: Criar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getCustomers
      tags: [Customers]
      summary: Listar todos os clientes
      parameters:
        - name: document
          in: query
          description: Filtrar pelo CPF ou CNPJ; pontuacao e ignorada.
          schema:
            type: string
      responses:
        <<: *protectedResponses
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Customer"
  /customers/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getCustomersId
      tags: [Customers]
      summary: Buscar cliente por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putCustomersId
      tags: [Customers]
      summary: Atualizar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteCustomersId
      tags: [Customers]
      summary: Deletar cliente
      responses:
        <<: *protectedResponses
        "204":
          description: Cliente deletado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Vehicles ──────────────────────────────────────────
  /vehicles:
    post:
      operationId: postVehicles
      tags: [Vehicles]
      summary: Criar veiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Veiculo criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getVehicles
      tags: [Vehicles]
      summary: Listar todos os veiculos
      parameters:
        - name: customer_id
          in: query
          description: Filtrar por cliente (UUID). O alias legado `customerId` tambem e aceito temporariamente.
          schema:
            type: string
            format: uuid
      responses:
        <<: *protectedResponses
        "200":
          description: Lista de veiculos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getVehiclesId
      tags: [Vehicles]
      summary: Buscar veiculo por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Veiculo encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putVehiclesId
      tags: [Vehicles]
      summary: Atualizar veiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Veiculo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteVehiclesId
      tags: [Vehicles]
      summary: Deletar veiculo
      responses:
        <<: *protectedResponses
        "204":
          description: Veiculo deletado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Supplies ──────────────────────────────────────────
  /supplies/pending-purchases:
    get:
      operationId: getSuppliesPendingPurchases
      tags: [Supplies]
      summary: Listar compras pendentes (insumos em falta para servicos aprovados)
      description: Retorna todos os servicos aprovados que possuem insumos com quantidade necessaria maior que o estoque disponivel.
      responses:
        <<: *protectedResponses
        "200":
          description: Lista de alertas de compra
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/SupplyShortageAlert"
  /supplies:
    post:
      operationId: postSupplies
      tags: [Supplies]
      summary: Criar insumo/peca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Insumo criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getSupplies
      tags: [Supplies]
      summary: Listar todos os insumos
      responses:
        <<: *protectedResponses
        "200":
          description: Lista de insumos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Supply"
  /supplies/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getSuppliesId
      tags: [Supplies]
      summary: Buscar insumo por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Insumo encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putSuppliesId
      tags: [Supplies]
      summary: Atualizar insumo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplyRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Insumo atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteSuppliesId
      tags: [Supplies]
      summary: Deletar insumo
      responses:
        <<: *protectedResponses
        "204":
          description: Insumo deletado
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  # ── Services ──────────────────────────────────────────
  /services:
    post:
      operationId: postServices
      tags: [Services]
      summary: Criar servico da oficina
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopServiceCreateRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Servico criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getServices
      tags: [Services]
      summary: Listar servicos da oficina
      parameters:
        - name: title
          in: query
          description: Busca parcial por titulo (case insensitive)
          schema:
            type: string
        - name: active
          in: query
          description: Filtrar por status ativo
          schema:
            type: boolean
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        <<: *protectedResponses
        "200":
          description: Lista paginada de servicos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /services/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getServicesId
      tags: [Services]
      summary: Buscar servico por ID
      responses:
        <<: *protectedResponses
        "200":
          description: Servico encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putServicesId
      tags: [Services]
      summary: Atualizar servico (parcial)
      description: Apenas os campos fornecidos sao alterados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkshopServiceUpdateRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: Servico atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: deleteServicesId
      tags: [Services]
      summary: Deletar ou desativar servico
      description: |
        Hard delete se nao ha vinculos com OS.
        Se ha vinculos, desativa (soft delete) e retorna 200.
      responses:
        <<: *protectedResponses
        "204":
          description: Servico deletado (hard delete)
        "200":
          description: Servico desativado (soft delete)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: service has existing work order links and was deactivated
                  service:
                    $ref: "#/components/schemas/WorkshopServiceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /services/avg-execution-time:
    get:
      operationId: getServicesAvgExecutionTime
      tags: [Reports]
      summary: Tempo medio de execucao por servico
      description: |
        Retorna o tempo medio real de execucao para cada tipo de servico,
        calculado a partir de servicos finalizados em ordens de servico.
      parameters:
        - name: from
          in: query
          description: Data inicio (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: to
          in: query
          description: Data fim (YYYY-MM-DD)
          schema:
            type: string
            format: date
            example: "2026-12-31"
        - name: technician_id
          in: query
          description: Filtrar pelo tecnico atribuido (UUID). O alias legado `technicianId` tambem e aceito temporariamente.
          schema:
            type: string
            format: uuid
      responses:
        <<: *protectedResponses
        "200":
          description: Lista com tempos medios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AvgExecutionTimeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Work Orders ───────────────────────────────────────
  /work-orders:
    post:
      operationId: postWorkOrders
      tags: [Work Orders]
      summary: Criar ordem de servico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderCreateRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: OS criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: getWorkOrders
      tags: [Work Orders]
      summary: Listar ordens de servico
      description: |
        Lista paginada de OS operacionais. Regras de exclusao e ordenacao:
        - FINALIZADA e ENTREGUE sao SEMPRE excluidas, inclusive quando o filtro `status`
          e informado (o filtro explicito NAO contorna a exclusao). Filtrar por um desses
          status retorna lista vazia (total = 0).
        - CANCELADA fica oculta na listagem padrao, mas e retornada quando o filtro
          explicito `status=CANCELADA` e informado (consulta/auditoria).
        - APROVADO permanece na listagem (OS ativa aguardando execucao); como nao esta
          entre os status priorizados, ordena depois deles.
        - Ordenacao por prioridade de status: EM_EXECUCAO > AGUARDANDO_APROVACAO >
          EM_DIAGNOSTICO > RECEBIDA (demais status ordenam por ultimo).
        - Dentro do mesmo status, mais antigas primeiro (received_at ASC).
        - Paginacao (`total`, `total_pages`) considera as exclusoes acima.
      parameters:
        - name: status
          in: query
          description: >-
            Filtrar por status. FINALIZADA e ENTREGUE permanecem sempre excluidas
            (filtrar por elas retorna vazio). CANCELADA so aparece quando filtrada
            explicitamente.
          schema:
            $ref: "#/components/schemas/WorkOrderStatus"
        - name: customer_id
          in: query
          description: Filtrar por cliente (UUID). O alias legado `customerId` tambem e aceito temporariamente.
          schema:
            type: string
            format: uuid
        - name: vehicle_id
          in: query
          description: Filtrar por veiculo (UUID). O alias legado `vehicleId` tambem e aceito temporariamente.
          schema:
            type: string
            format: uuid
        - name: from
          in: query
          description: Data inicio (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Data fim (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        <<: *protectedResponses
        "200":
          description: Lista paginada de OS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrderListResponse"
        "401":
          description: Nao autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /work-orders/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    get:
      operationId: getWorkOrdersId
      tags: [Work Orders]
      summary: Buscar OS por ID
      responses:
        <<: *protectedResponses
        "200":
          description: OS encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: putWorkOrdersId
      tags: [Work Orders]
      summary: Atualizar OS (campos e/ou status)
      description: |
        Se o campo `status` for enviado, a transicao de status e processada via maquina de estados.
        Ao transitar para AGUARDANDO_APROVACAO, o orcamento e enviado por e-mail.
        Na geracao do orcamento, quando houver insumo com quantidade necessaria maior que o estoque,
        o prazo estimado do servico no orcamento recebe acrescimo padrao de 2 dias.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderUpdateRequest"
      responses:
        <<: *protectedResponses
        "200":
          description: OS atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Transicao de status invalida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /work-orders/{id}/services:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    post:
      operationId: postWorkOrdersIdServices
      tags: [Work Orders]
      summary: Adicionar servicos a uma OS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/AddServiceRequest"
      responses:
        <<: *protectedResponses
        "201":
          description: Servicos adicionados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrderService"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          description: Status da OS nao permite adicionar servicos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /work-orders/{id}/services/{wosId}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: wosId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID do servico na OS (work_order_service)
    delete:
      operationId: deleteWorkOrdersIdServicesWosId
      tags: [Work Orders]
      summary: Remover servico de uma OS
      responses:
        <<: *protectedResponses
        "204":
          description: Servico removido
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Status da OS nao permite remocao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /work-orders/{id}/services/{wosId}/supplies/{supplyId}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
      - name: wosId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: supplyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteWorkOrdersIdServicesWosIdSuppliesSupplyId
      tags: [Work Orders]
      summary: Remover insumo de um servico da OS
      responses:
        <<: *protectedResponses
        "204":
          description: Insumo removido
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          description: Status nao permite remocao
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ── Public ────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      description: Numero da pagina (default 1)
      schema:
        type: integer
        minimum: 1
        default: 1
...
  responses:
    BadRequest:
      description: Requisicao invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid id"
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [admin, employee]
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, employee]

    # ── Customer ──────────────────────────────────────
...